requirements.txt
Flask==3.1.2
openai==2.8.0
httpx==0.28.1
python-docx==1.2.0
markdown==3.10
